Fixed foreign key problem for south migrations
