+++ pybb/models/__init__.py
@@ -1,12 +1,11 @@
 from pybb.models.attachment import Attachment
 from pybb.models.category import Category
 from pybb.models.forum import Forum
-from pybb.models.poll_answer import PollAnswer, PollAnswerUser
 from pybb.models.post import Post
 from pybb.models.profile import PybbProfile, Profile
 from pybb.models.read_trackers import ForumReadTracker, TopicReadTracker
 from pybb.models.renderable import RenderableItem
-from pybb.models.topic import Topic
+from pybb.models.topic import Topic, PollAnswer, PollAnswerUser
 
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext as _
